requirements.txt
appdirs==1.4.4
argcomplete==3.2.1
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
py4j==0.10.9.7
PyJWT==2.8.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
python-barcode==0.15.1
python-escpos==3.1
pytz==2023.3.post1
pyusb==1.2.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
WMI==1.5.1
